feat: add cluster support
