+++ .env
@@ -1,2 +1,1 @@
-HOST="localhost"
 PORT="4000"
