+++ src/index.ts
@@ -1,59 +1,25 @@
 import "dotenv/config";
-import { createServer } from "node:http";
 import { argv } from "node:process";
-import { read } from "./controllers/read";
-import { create } from "./controllers/create";
-import { update } from "./controllers/update";
-import { remove } from "./controllers/remove";
+import { availableParallelism } from "node:os";
+import cluster from "node:cluster";
+import { join } from "node:path";
 
 const PORT = Number(process.env.PORT) || 4000;
-const HOST = process.env.HOST || "localhost";
 
 const args = argv.slice(2);
 
 const isMulti = !!args.find((arg) => arg === "--multi");
 
-if (!isMulti) {
-  const server = createServer((request, response) => {
-    try {
-      const { url, method } = request;
+const cores = availableParallelism();
 
-      if (!url?.startsWith("/api/users")) {
-        response.writeHead(404, "Not Found");
-        response.end(
-          JSON.stringify({ message: "Invalid API Endpoint Route" })
-        );
-      } else {
-        switch (method) {
-          case "GET":
-            read(request, response);
-            break;
-          case "POST":
-            create(request, response);
-            break;
-          case "PUT":
-            update(request, response);
-            break;
-          case "DELETE":
-            remove(request, response);
-            break;
-          default:
-            response.writeHead(405, "Invalid Request");
-            response.end(
-              JSON.stringify({ message: "Invalid API request method" })
-            );
-            break;
-        }
-      }
-    } catch (error) {
-      response.writeHead(500, "Server Error");
-      response.end(
-        JSON.stringify({ message: "Internal Server Error" })
-      );
-    }
-  });
+cluster.setupPrimary({
+  exec: join(__dirname, "server"),
+});
 
-  server.listen(PORT, HOST, () => {
-    console.log(`Listen on port ${PORT}`);
+if (isMulti) {
+  Array.from({ length: cores }).forEach((_, index) => {
+    cluster.fork({ PORT: PORT + index });
   });
+} else {
+  cluster.fork();
 }
